Schedule on need
when a truck becomes idle or a request is received

+++ server/package_test.go
@@ -11,7 +11,7 @@ import (
 func TestPackageIdReqs(t *testing.T) {
 	resp, err := server.PackageIdReqs([]*bridge.Package{
 		{
-			WarehouseId: proto.Int64(1),
+			WarehouseId: proto.Int32(1),
 			X:           proto.Int32(233),
 			Y:           proto.Int32(666),
 		},
